requirements.txt
# Core
streamlit==1.40.1
python-dotenv==1.0.1

# LLM & Embeddings
groq==0.32.0
langchain==0.3.13
langchain-community==0.3.13
google-generativeai==0.7.2  # Add this for Gemini
langchain==0.3.13
langchain-community==0.3.13
# Embeddings
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# Vector Store
faiss-cpu==1.9.0

# PDF Processing
PyMuPDF==1.24.14

# Utilities
requests==2.32.3
